add mocha and first unit test
This commit sets up the unit testing framework mocha and adds a single
test for the Player constructor.

It was necessary to pull the Player object out of app.js to avoid a local
server being started upon running mocha.

There is also some nuance to the initial test in that it avoids an
infinite loop that exists inside of the Player constructor.